Added support for standardized indexer cron job
Every application is going to need to run the indexer at some point.
This change standardizes how it's run, and lets the user define when to
run it.

Basically it's wrapping the cron type, and passing in whatever values
are in configuration[:sphinx][:index_cron]. For example, it can be
configured like:

    # in app/manifests/application_manifest.rb
    configure :sphinx => {
      :version => '0.9.9' # 0.9.8.1 is the default
      # run the cron every hour at 9 (no default)
      :index_cron => {
        :minute => 9
      }
    }
    # or in config/moonshine.yml
    :sphinx:
      :version: 0.9.9
      # run the cron every hour at 9 (no default)
      :index_cron:
        :minute: 9